+++ CHANGELOG.md
@@ -10,12 +10,14 @@
 - Rewriting encoding/decoding to support forward and backward compitibility.
 - Support `zstd`, `brotli`, `lz4`, and `snappy` for compressing.
 - Support `crc32`, `xxhash64`, `xxhash32`, `xxhash3`, `murmur3` for checksuming.
+- Unify returned error, all exported APIs return `Error` on `Result::Err`.
 
 ### Breakage
 
 - Remove `native-tls` supports
 - Remove `s2n-quic` supports
 - Remove `Wire` trait to simplify `Transport` trait
+- Remove `JoinError`, add an new `Error::Multiple` variant
 
 ### Testing